seagent/requirements.txt
openai==1.21.2
pydantic==2.7.0
instructor==1.2.0
httpx==0.27.0
pandas==2.2.2
openpyxl==3.1.5
py2neo==2021.2.4
streamlit==1.37.1
streamlit-authenticator==0.3.2
neo4j==5.2.1
python-dotenv==1.0.1
pypict==0.3.0
mysql-connector-python==8.4.0
jsonschema==4.21.1

